+++ NEWS.md
@@ -3,6 +3,7 @@
 ## enhancements
 
 - `spatUnit()<-` and `featType()<-` `list` methods
+- `set_default_spat_unit()` and `set_default_feat_type()` now look for defaults when given `NA_character_` inputs as well.
 
 # GiottoClass 0.4.4 (2024/11/14)